Put Scanner into separate class
