+++ src/Controller/ToDoController.java
@@ -2,18 +2,16 @@ package Controller;
 import Model.Task;
 import Model.TasksList;
 import View.Printer;
-import View.Reader;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Scanner;
 
 public class ToDoController {
     private TasksList collection;
     private boolean programIsRunning = true;
-    private Scanner userInput = new Scanner(System.in);
-    private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
+   // private Scanner userInput = new Scanner(System.in);
+   private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
 
     public ToDoController(){
         this.collection = new TasksList();
@@ -49,7 +47,7 @@ public class ToDoController {
     {
         System.out.println("we will remove");
         collection.showAllTaskByDate();
-        int index = userInput.nextInt();//Reader.readInt();
+        int index = Reader.readInt(); //userInput.nextInt();
         //userInput.nextLine();
         //TODO check that the index is int, check that this index is in the list
 
@@ -62,7 +60,7 @@ public class ToDoController {
     }
     private void showTaskList() {
         Printer.taskListMenu();
-        String menuOption = userInput.nextLine();
+        String menuOption = Reader.readLine(); //userInput.nextLine();
         // validate user input
         switch (menuOption) {
             case "1":
@@ -85,7 +83,7 @@ public class ToDoController {
 
     private void showTaskListByProjectName() {
         System.out.println("Please insert the name of the project");
-        String projectName = userInput.nextLine();
+        String projectName = Reader.readLine();//userInput.nextLine();
         if (collection.getTaskListSize() == 0) {
             System.out.println("Your task list is empty. Please create and add a task");
         } else if (collection.containsProjectName(projectName)) {
@@ -96,8 +94,8 @@ public class ToDoController {
     }
         private void showTaskListByDate(){
             System.out.println("Please insert the date");
-            String deadline = userInput.nextLine();
-            LocalDate date = LocalDate.parse(deadline, formatter);//Validator.validatDate();
+            String deadline = Reader.readLine();//userInput.nextLine();
+            LocalDate date = Validator.validatDate();//LocalDate.parse(deadline, formatter);//Validator.validatDate();
             if (collection.getTaskListSize() == 0) {
                 System.out.println("Your task list is empty. Please create and add a task");
             } else if (collection.containsDateName(date)) {
@@ -109,17 +107,17 @@ public class ToDoController {
 
     private void insertDataForTask(ArrayList<Task> tasksList){
         Printer.getTaskTitle();
-        String taskName = userInput.nextLine();
+        String taskName = Reader.readLine(); //userInput.nextLine();
 
         Printer.getTaskDate();
         //String dueDate = userInput.nextLine();
         LocalDate deadline = Validator.validatDate();// LocalDate.parse(dueDate, formatter);
 
         Printer.getTaskProject();
-        String projectName = userInput.nextLine();
+        String projectName = Reader.readLine(); //userInput.nextLine();
 
         Printer.getTaskStatus();
-        String status = userInput.nextLine();
+        String status = Reader.readLine(); //userInput.nextLine();
 
         //Here we add a created task to a created taskList
         tasksList.add(new Task(taskName, deadline, formatter, projectName, status));
